Rozwój projektu:
1. Poprawa modeli
   - Usunięcie modelu DaneUzytkownika
   - Produkt - zmiana na liczby całkowite cen
   - Uzytkownik - przeniesienie danych z modelu DaneUzytkownika do        Uzytkownik oraz dodatnie numeru telefonu
   - Zamowienia - dodanie danych odbiorcy i metody płatności
2. Poprawa widoków o zaktualizowane modele
3. Dodanie nowych funkcji do kontrolera i zaktualizowanie już istniejących.
4. Wykończenie Index'u o wygląd CSS .

+++ ProjektStrona_EG_IG/Models/Produkt.cs
@@ -18,8 +18,8 @@ namespace ProjektStrona_EG_IG.Models
         [MaxLength(100)]
         public string Opis { get; set; }
 
-        [Range(0.0, double.MaxValue, ErrorMessage = "Cena nie może być mniejsza niż 0.")]
-        public decimal Cena { get; set; }
+        [Range(0, int.MaxValue, ErrorMessage = "Cena nie może być mniejsza niż 0.")]
+        public int Cena { get; set; }
 
         [Range(0, int.MaxValue, ErrorMessage = "Ilość dostępnych produktów nie może być mniejsza niż 0.")]
         public int IloscDostepna { get; set; }
